requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
SQLAlchemy==1.4.37
psycopg2==2.9.3
modelyst-sqlmodel==0.0.9
alembic==1.8.0
inflection==0.5.1
sqladmin==0.1.10
GeoAlchemy2==0.12.1
strawberry-graphql==0.114.3
Shapely==1.8.2
